LayoutTests/webgl/query-context-lost-restore.html

<!DOCTYPE html><!-- webkit-test-runner [ WebGLTimerQueriesEnabled=true ] -->
<html>
<head>
<meta charset="utf-8">
<link rel="stylesheet" href="resources/webgl_test_files/resources/js-test-style.css"/>
<script src="resources/webgl_test_files/js/js-test-pre.js"></script>
<script src="resources/webgl_test_files/js/webgl-test-utils.js"></script>
</head>
<body onload="runTest()">
<div id="description"></div>
<div id="console"></div>
<script>
"use strict";
description("Tests that active query state is properly reset after context loss: queries can be begun after restore and active query objects are collectible.");

const wtu = WebGLTestUtils;

var gl;

function waitForEvent(target, eventName, handler) {
    return new Promise((resolve, reject) => {
        setTimeout(() => reject(new Error("Timed out waiting for " + eventName)), 5000);
        target.addEventListener(eventName, (e) => {
            if (handler)
                handler(e);
            resolve(e);
        }, { once: true });
    });
}

async function testWebGL1TimerQuery() {
    debug("Running test: contextType: webgl, queryTarget: TIME_ELAPSED_EXT");

    const canvas = document.createElement("canvas");
    canvas.width = 1;
    canvas.height = 1;
    gl = wtu.create3DContext(canvas);
    if (!gl) {
        testFailed("Could not create webgl context");
        return;
    }

    const WEBGL_lose_context = wtu.getExtensionWithKnownPrefixes(gl, "WEBGL_lose_context");
    if (!WEBGL_lose_context) {
        testPassed("WEBGL_lose_context not available, skipping");
        return;
    }

    let ext = gl.getExtension("EXT_disjoint_timer_query");
    if (!ext) {
        testPassed("EXT_disjoint_timer_query not available, skipping");
        return;
    }

    let query = ext.createQueryEXT();
    ext.beginQueryEXT(ext.TIME_ELAPSED_EXT, query);
    wtu.glErrorShouldBe(gl, gl.NO_ERROR, "beginQueryEXT should succeed");

    const contextLostPromise = waitForEvent(canvas, "webglcontextlost", (e) => e.preventDefault());
    WEBGL_lose_context.loseContext();
    await contextLostPromise;
    shouldBeTrue("gl.isContextLost()");

    const contextRestoredPromise = waitForEvent(canvas, "webglcontextrestored");
    await new Promise(resolve => setTimeout(resolve, 0));
    WEBGL_lose_context.restoreContext();
    await contextRestoredPromise;
    shouldBeFalse("gl.isContextLost()");

    ext = gl.getExtension("EXT_disjoint_timer_query");
    if (!ext) {
        testFailed("Could not re-enable EXT_disjoint_timer_query after restore");
        return;
    }

    const newQuery = ext.createQueryEXT();
    ext.beginQueryEXT(ext.TIME_ELAPSED_EXT, newQuery);
    wtu.glErrorShouldBe(gl, gl.NO_ERROR, "beginQueryEXT with new query should succeed after restore");

    ext.endQueryEXT(ext.TIME_ELAPSED_EXT);
    wtu.glErrorShouldBe(gl, gl.NO_ERROR, "endQueryEXT should succeed");
}

async function testWebGL2Query(queryTarget, queryTargetName, requiresExtension) {
    debug(`Running test: contextType: webgl2, queryTarget: ${queryTargetName}`);

    const canvas = document.createElement("canvas");
    canvas.width = 1;
    canvas.height = 1;
    gl = wtu.create3DContext(canvas, undefined, 2);
    if (!gl) {
        testFailed("Could not create webgl2 context");
        return;
    }

    const WEBGL_lose_context = wtu.getExtensionWithKnownPrefixes(gl, "WEBGL_lose_context");
    if (!WEBGL_lose_context) {
        testPassed("WEBGL_lose_context not available, skipping");
        return;
    }

    if (requiresExtension) {
        const ext = gl.getExtension(requiresExtension);
        if (!ext) {
            testPassed(`${requiresExtension} not available, skipping`);
            return;
        }
    }

    let query = gl.createQuery();
    gl.beginQuery(queryTarget, query);
    wtu.glErrorShouldBe(gl, gl.NO_ERROR, `beginQuery(${queryTargetName}) should succeed`);

    const contextLostPromise = waitForEvent(canvas, "webglcontextlost", (e) => e.preventDefault());
    WEBGL_lose_context.loseContext();
    await contextLostPromise;
    shouldBeTrue("gl.isContextLost()");

    const contextRestoredPromise = waitForEvent(canvas, "webglcontextrestored");
    await new Promise(resolve => setTimeout(resolve, 0));
    WEBGL_lose_context.restoreContext();
    await contextRestoredPromise;
    shouldBeFalse("gl.isContextLost()");

    if (requiresExtension) {
        const ext = gl.getExtension(requiresExtension);
        if (!ext) {
            testFailed(`Could not re-enable ${requiresExtension} after restore`);
            return;
        }
    }

    const newQuery = gl.createQuery();
    gl.beginQuery(queryTarget, newQuery);
    wtu.glErrorShouldBe(gl, gl.NO_ERROR, `beginQuery(${queryTargetName}) with new query should succeed after restore`);

    gl.endQuery(queryTarget);
    wtu.glErrorShouldBe(gl, gl.NO_ERROR, `endQuery(${queryTargetName}) should succeed`);
}

var WebGL2QueryTargets = {
    ANY_SAMPLES_PASSED                      : 0x8C2F,
    TRANSFORM_FEEDBACK_PRIMITIVES_WRITTEN   : 0x8C88,
    TIME_ELAPSED_EXT                        : 0x88BF,
};

async function runTest() {
    await testWebGL1TimerQuery();
    await testWebGL2Query(WebGL2QueryTargets.ANY_SAMPLES_PASSED, "ANY_SAMPLES_PASSED", null);
    await testWebGL2Query(WebGL2QueryTargets.TRANSFORM_FEEDBACK_PRIMITIVES_WRITTEN, "TRANSFORM_FEEDBACK_PRIMITIVES_WRITTEN", null);
    await testWebGL2Query(WebGL2QueryTargets.TIME_ELAPSED_EXT, "TIME_ELAPSED_EXT", "EXT_disjoint_timer_query_webgl2");
    finishTest();
}
</script>
</body>
</html>
