LayoutTests/webgl/query-context-lost-restore-expected.txt

Tests that active query state is properly reset after context loss: queries can be begun after restore and active query objects are collectible.

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".

TEST COMPLETE: 21 PASS, 0 FAIL

Running test: contextType: webgl, queryTarget: TIME_ELAPSED_EXT
PASS getError was expected value: NO_ERROR : beginQueryEXT should succeed
PASS gl.isContextLost() is true
PASS gl.isContextLost() is false
PASS getError was expected value: NO_ERROR : beginQueryEXT with new query should succeed after restore
PASS getError was expected value: NO_ERROR : endQueryEXT should succeed
Running test: contextType: webgl2, queryTarget: ANY_SAMPLES_PASSED
PASS getError was expected value: NO_ERROR : beginQuery(ANY_SAMPLES_PASSED) should succeed
PASS gl.isContextLost() is true
PASS gl.isContextLost() is false
PASS getError was expected value: NO_ERROR : beginQuery(ANY_SAMPLES_PASSED) with new query should succeed after restore
PASS getError was expected value: NO_ERROR : endQuery(ANY_SAMPLES_PASSED) should succeed
Running test: contextType: webgl2, queryTarget: TRANSFORM_FEEDBACK_PRIMITIVES_WRITTEN
PASS getError was expected value: NO_ERROR : beginQuery(TRANSFORM_FEEDBACK_PRIMITIVES_WRITTEN) should succeed
PASS gl.isContextLost() is true
PASS gl.isContextLost() is false
PASS getError was expected value: NO_ERROR : beginQuery(TRANSFORM_FEEDBACK_PRIMITIVES_WRITTEN) with new query should succeed after restore
PASS getError was expected value: NO_ERROR : endQuery(TRANSFORM_FEEDBACK_PRIMITIVES_WRITTEN) should succeed
Running test: contextType: webgl2, queryTarget: TIME_ELAPSED_EXT
PASS getError was expected value: NO_ERROR : beginQuery(TIME_ELAPSED_EXT) should succeed
PASS gl.isContextLost() is true
PASS gl.isContextLost() is false
PASS getError was expected value: NO_ERROR : beginQuery(TIME_ELAPSED_EXT) with new query should succeed after restore
PASS getError was expected value: NO_ERROR : endQuery(TIME_ELAPSED_EXT) should succeed
PASS successfullyParsed is true

TEST COMPLETE
